* [jsfm] move common methods into operation.js
